Add tree builders module
  - Add builder function to package API
  - Switch to v3.0.0b2 version
