+++ doc/conf.py
@@ -31,7 +31,7 @@ author = 'Davide Brunato'
 # The short X.Y version
 version = '3.0'
 # The full version, including alpha/beta/rc tags
-release = '3.0.0b1'
+release = '3.0.0b2'
 
 # -- General configuration ---------------------------------------------------
